CODECRAFT-35 fix most of memory-mgr tests (#10)
## Fixes

Fix ctest regex to detect fatal errors as errors
Add missing includes where needed to fix clangd errors in files
(compilation mainly worked without these fixes)
Re-enable allocator_decorator constructor with argument as it is needed
for decorator managers. Added CODECRAFT-89 to improve it with use of
enable tag for pimpl constructible.
Add `pthread)mutexattr_init` call to fix mutex init on linux. Otherwise
it was causing random crashes due to having garbage attributes. Issue
never occurred on mac, but the same fix was added for symmetry. Init all
synchronization structs using `{}` to make sure they are zero init on
all platforms in all build configs.
Inline `ulong_cast` to break include cycle in `win32.h`
Use size_t and ptrdiff_t as default for non cross-arch compatible types
to match C++ default.
Move typecast to unsigned into `math::lowest_bit` instead of using
assert. This way it always work no matter if argument is signed or
unsigned. Bits are bits after all.
Fix `size_tracking` and `new_helper` to use `portable_size_t` instead of
mismatching `size_t` and `ulonglong` that was causing crashes on Win64
and linux.
Add missing/remove redundant `typename`/`template` keywords (simply
following compiler errors/clangd autofixes)
Remove redundant types from the typelists for testing
Increase memory_size for some tests as too small size causes build
errors (yet to figure that out, but its not really important for now)
Fix dark magic cast of test_new to offset vmptr_size to actual object
data.
Remove invalid tests for `shared_ptr` that pass too many arguments into
`allocate_shared` and `make_shared` which somehow used to work for non
variadic templates versions

### TODO
`pool/test_pool_alloc.cpp` still needs to be fixed
`pool/time_pool_alloc.cpp` - is actually a performance test and need to
be moved, still needs fixes as well.
`named_mutex` is not implemented for non Windows platforms and thus
corresponding tests are still disabled

## Other changes
Build RelWithDebInfo to have better crash stacks in CI
Add more recommended extensions that are useful for project development

Use separate build folders per build type to reduce rebuilds
Update git ignores to reflect settings updates

Add add_elevated_manifest to CMake which is useful for debugging memory
manager tests on Windows when UAC is enabled and VSCode isn't running
under elevated privileges. It ended up not used, because debugging
simply crashes to start if VSCode is not elevated. Better solution would
be to not use global segment for tests. That would allow proper
parallelization of tests as well. Tracked in CODECRAFT-93

Add debug_break that can be enabled in assert for easier debugging.
CODECRAFT-91 tracks auto enable under debugger

+++ code/memory-mgr/include/memory-mgr/detail/compatibility_types.h
@@ -36,8 +36,8 @@ namespace memory_mgr
 		typedef uint64_t portable_size_t;
 		typedef int64_t portable_difference_type;
 #else
-		typedef ulong portable_size_t;
-		typedef long portable_difference_type;
+		typedef size_t portable_size_t;
+		typedef ptrdiff_t portable_difference_type;
 #endif
 	}
 }
